Fix Zero From Popping Up When using Negative Btn
When the user would press the negative button AFTER they used the clear
button while there is no number inputted yet, the calculator would
automatically put a zero in the display. This was because I set the
displayValue to 0 when pressing the clear button instead of setting it
to a blank string.

Also ignore the changes to the expression object and equals button
eventListener. All I did was change how 2 inputs are evaluated. Still trying
to figure out how to make the evaluation work when chaining multiple
operations