Fix Delete Button After an Expression is Eval
Whenever an expression was solved after using the equal button, the
delete button would not act accordingly. Either the entire number would
be deleted, or seen to work on the frontend, but in the backend
would not. Chaining more expressions after the equal button was used did
not properly calculate. That has now been fixed.

How the calculator previously worked:
1+2+7 = 10
10 => 1 *press the delete button once, 10 now becomes 1 or fully deletes*
1+2 = incorrect calculation

How the calculator works now:
1+2+7 = 10
10 => 1 *user presses the delete button once, 10 now becomes 1*
1+2 = 3 (correct calculation)