Add the Function determineInputBehavior()
Moved the repeating if else statements that were in the operations btn
function moved them all into a single function. This makes the code more
readable, and easier to make changes. Instead of having to make sure
every location of these if else statements was the same, I will only have
to worry about editing a single function.